+++ test/bs_flutter_utils_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:bs_flutter_utils/bs_flutter_utils.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('bs_flutter_utils');
@@ -15,9 +14,5 @@ void main() {
 
   tearDown(() {
     channel.setMockMethodCallHandler(null);
-  });
-
-  test('getPlatformVersion', () async {
-    expect(await BsFlutterUtils.platformVersion, '42');
   });
 }